backend/pyproject.toml
[tool.poetry]
name = "langgraph-livekit-agents"
version = "0.0.1"
description = "LiveKit LangGraph adapter"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.3.15"
langchain-openai = "^0.3.1"
langgraph = "^0.2.66"
langgraph-cli = "^0.1.55"
langgraph-api = "^0.2.86"
livekit-agents = "^0.12.9"
livekit-plugins-deepgram = "^0.6.16"
livekit-plugins-openai = "^0.10.16"
livekit-plugins-silero = "^0.7.4"
livekit-plugins-elevenlabs = "^0.7.4"
python-dotenv = "^1.0.1"
fastapi = "^0.104.1"
uvicorn = "^0.26.0"
livekit = "^0.12.0"
httpx = "^0.25.0"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
